Fixes date display bug.
Items that are undated, or have non-date values for DC:Date, were
displaying 'December 31,1969' as their date. This commit adds a helper
function to check whether a given string is a valid date, and implements
that helper on the items/show and items/browse pages.

This also moves the date display on the items/show to the top of the
page, as an eyebrow on the item title, instead of in the table of other
item metadata.